requirements.txt
# Base de datos
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Procesamiento de datos
pandas==2.1.4
numpy==1.26.2

# Excel
openpyxl==3.1.2
XlsxWriter==3.1.9
pyxlsb==1.0.10

# Dashboard Web
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# API (opcional)
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3

# Scheduler
APScheduler==3.10.4
python-crontab==3.0.0

# Utilidades
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0

# Logging y monitoreo
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Desarrollo
black==23.12.1
flake8==7.0.0
mypy==1.7.1
